requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.4.3-cp38-cp38-win32.whl
allure-pytest